% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-cols.R
\name{add_rnk_col}
\alias{add_rnk_col}
\title{\code{dplyr::mutate(... dplyr::row_number(dplyr::desc(.)))}}
\usage{
add_rnk_col(data, col, col_out = "rnk", arrange = TRUE, pretty = TRUE)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{col}{A character for the SE version; a symbol for NSE version.
The name of column in \code{data} on which to perform operations.}

\item{col_out}{Name of the new column to create, as a string or symbol.}

\item{arrange}{If \code{TRUE}, will arrange \code{data} according to \code{col_out}.}

\item{pretty}{If \code{TRUE}, will re-order columns such that \code{col_out} is the first column.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Shorthand for \code{{dplyr}} functions called consecutively.
}
\section{Borrowing from the \code{{tidyverse}}}{

Documentation for \code{col} and \code{col_out} is modified
\code{tidyr::gather()}'s documentation of \code{key} and \code{value}.
}

\seealso{
Other dplyr-cols: \code{\link{select_one_of}}
}
