% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-combos.R
\name{arrange_distinctly}
\alias{arrange_distinctly}
\title{\code{dplyr::distinct()} + \code{dplyr::arrange()}}
\usage{
arrange_distinctly(data, ..., .keep_all = FALSE, .by_group = FALSE)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{...}{Additional arguments passed to internal functions.}

\item{.keep_all}{If \code{TRUE}, keep all variables in \code{.data}.
If a combination of \code{...} is not distinct, this keeps the
first row of values.}

\item{.by_group}{If \code{TRUE}, will sort first by grouping variable. Applies to
grouped data frames only.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Shorthand for \code{{dplyr}} functions called consecutively.
}
\section{Borrowing from the \code{{tidyverse}}}{

The logic for handling \code{...} is borrowed from a reivew of the internals for
\code{tidyr::unite()} and \code{tidyr::gather()}.
}

\seealso{
Other dplyr-combos: \code{\link{count_arrange}},
  \code{\link{pull_distinctly}}
}
