% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-combos.R
\name{count_arrange_at}
\alias{count_arrange_at}
\title{\code{dplyr::count()} + \code{dplyr::arrange()}}
\usage{
count_arrange_at(data = NULL, ...)
}
\arguments{
\item{data}{data.frame.}

\item{...}{dots. Bare names of columns in \code{data} on which to perform operations.}
}
\value{
data.frame.
}
\description{
Shorthand for \code{dplyr} functions called consecutively.
}
\details{
Note that arrangeing is performed on the column used to compute
the value for \code{dplyr::count()}, which is differrent than simply
calling \code{dplyr::count(..., sort = TRUE)} (which sorts by the output column \code{n}).
}
