% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-combos.R
\name{pull_distinctly}
\alias{pull_distinctly}
\title{\code{dplyr::pull()} + \code{unique()} + \code{sort()}}
\usage{
pull_distinctly(.data, var = -1, ..., decreasing = FALSE)
}
\arguments{
\item{.data}{A table of data}

\item{var}{A variable specified as:
\itemize{
\item a literal variable name
\item a positive integer, giving the position counting from the left
\item a negative integer, giving the position counting from the right.
}

The default returns the last column (on the assumption that's the
column you've created most recently).

This argument is taken by expression and supports
\link[rlang:quasiquotation]{quasiquotation} (you can unquote column
names and column positions).}

\item{...}{Additional arguments passed to internal \code{{base}}/\code{{stats}} functions.}

\item{decreasing}{logical.  Should the sort be increasing or decreasing?
    For the \code{"radix"} method, this can be a vector of
    length equal to the number of arguments in \code{\dots}. For the
    other methods, it must be length one. Not available for partial sorting.}
}
\value{
A vector (of the same type as \code{var}).
}
\description{
Shorthand for \code{{dplyr}} functions called consecutively.
}
\details{
This may be slower than \code{dplyr:;distinct()} +
\code{dplyr::arrange()} + \code{dplyr::pull()} (which was the original implementation).
}
\seealso{
Other dplyr-combos: \code{\link{arrange_distinctly}},
  \code{\link{count_arrange}}
}
\concept{dplyr-combos}
