% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-combos.R
\name{rank_arrange_at}
\alias{rank_arrange_at}
\alias{rank_arrange}
\alias{count_arrange}
\title{\code{dplyr::mutate(... dplyr::row_number(dplyr::desc(.)))}}
\usage{
rank_arrange_at(data, col, col_out = "rnk", pretty = TRUE)

rank_arrange(data, col, ...)

count_arrange(data, ...)
}
\arguments{
\item{data}{data.frame.}

\item{col}{charactor for SE version; symbol for NSE version. Name of column in \code{data} on which to perform operations.}

\item{col_out}{charactor for SE version; symbol for NSE version. Name of column in returned data.frame
corresponding to rank.}

\item{pretty}{logical. Whether to re-arrange columns such that \code{col_out} is the first column.}

\item{...}{dots. For NSE version, passed to SE version.}
}
\value{
data.frame.
}
\description{
Shorthand for \code{dplyr} functions called consecutively.
}
\details{
None.
}
