% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-cols.R
\name{select_one_of}
\alias{select_one_of}
\title{\code{dplyr::select(data, tidyselect::one_of(cols), tidyselect::everything())}}
\usage{
select_one_of(data, cols, drop = FALSE)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{cols}{A character vector corresponding to the columns in \code{data} to be
selected by \code{dplyr::one_of()}.}

\item{drop}{If \code{FALSE}, does NOT drop all columns not specified by \code{cols}.
(i.e. This is like calling \code{tidyselect::everything()}).}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Shorthand for \code{{dplyr}} functions called consecutively.
}
\details{
There is probably a better way to implement this function internally.
Should review \code{tidyselect::one_of()} and \code{tidyr::separate()}'s \code{into} argument.
}
\seealso{
Other dplyr-cols: \code{\link{add_rnk_col}}
}
\concept{dplyr-cols}
