% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{summarise_join_stats}
\alias{summarise_join_stats}
\title{Calculate how "well" two data.frames joined}
\usage{
summarise_join_stats(data = NULL, x = NULL, y = NULL)
}
\arguments{
\item{data}{data.frame. This should already by the result of a join.}

\item{x, y}{Either characters or bare names of columns in \code{data} on which
to compute metrics.}
}
\value{
data.frame. One line summary of join metrics,
including the following columns:
\code{x}, \code{n_x}, \code{n_x_unjoined}, \code{x_in_y_pct} and their \code{y} analogues,
as well as \code{n_joined}.
}
\description{
Calculates the join count and percentage (by number of rows missing)
of a data.frame that is the results of a join.
}
\details{
This function is mostly useful after calling \code{dplyr::full_join()}
(or a similar function), although it can also be useful for \code{dplyr::left_join()}
or \code{dplyr::right_join()}.
}
\seealso{
\url{https://github.com/tidyverse/tidyr/blob/master/R/gather.R}
provides some insight for how to make the function accept either characters
or bare names for the column names.
}
