% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_stats}
\alias{summarise_stats}
\title{Summary statistics for a \code{data.frame}}
\usage{
summarise_stats(data, col, na.rm = TRUE, ..., tidy = FALSE,
  key = "stat", value = "value", convert = TRUE)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{col}{A character for the SE version; a symbol for NSE version.
The name of column in \code{data} on which to perform operations.}

\item{na.rm}{A logical value passed to \code{{base}}/\code{{stats}} functions used internally.}

\item{...}{Additional arguments passed to internal \code{{base}}/\code{{stats}} functions.}

\item{tidy}{A logical value indicating whether to put output in long (i.e. "tidy") format.}

\item{key, value, convert}{Parameters corresponding with those of the same name for \code{tidyr::gather()}.
For this function, \code{key} and \code{value} these should be characters.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
A customized summary of a data.frame.
}
\details{
The purpose of this function is equivalent to that of similar functions in
other packages, such as \code{skimr::skim()}.
This function outputs the following \code{n}, \code{mean}, \code{median}, \code{sd},
\code{min}, \code{max}, \code{zn1}, \code{zp1}, \code{q25}, \code{q75}, \code{q05}, \code{q95}.

The \code{_at} versions of this function are SE evaluation. (i.e. They take characters
as column names.) ~~The \code{_by} version(s) of this function allows
for groups to be specificed as an input, although this function will detect groups
and respect their integrity  (meaning that the \code{_by} version(s) are simply
provided as an alternative means).~~
}
\seealso{
\code{\link[skimr:skim]{skimr::skim()}}, \code{\link[tidyr:gather]{tidyr::gather()}}
}
