% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetidy-package.R
\docType{package}
\name{tetidy-package}
\alias{tetidy}
\alias{tetidy-package}
\title{tetidy: Tony's Personal Package}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

This is a package with functions for that Tony
    commonly uses for data munging workflows. These functions mostly
    depend on other from the tidyverse.
}
\author{
\strong{Maintainer}: Tony ElHabr \email{anthonyelhabr@gmail.com}

}
\keyword{internal}
