% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join_fuzzily}
\alias{join_fuzzily}
\title{Join two data.frames using fuzzy string matching.}
\usage{
join_fuzzily(x, y, mode = "inner", max_dist = 0, ...,
  cols_join = intersect(names(x), names(y)), copy = FALSE,
  suffix_x = "_x", suffix_y = "_y", suffix = c(suffix_x, suffix_y))
}
\arguments{
\item{x}{A tbl}

\item{y}{A tbl}

\item{mode}{One of "inner", "left", "right", "full" "semi", or "anti"}

\item{max_dist}{Maximum distance to use for joining}

\item{...}{Arguments passed on to \code{\link{stringdist}}}

\item{cols_join}{character. Names of columns in \code{x} and \code{y} to join upon.}

\item{copy}{logical. Whether or not to keep common columns}

\item{suffix_x, suffix_y}{For duplicate variables in \code{x} and \code{y}, these
are used for differentiating the two. Alternatively, \code{suffix} can be specified directly
(in the same manner for \code{suffix} in \code{dplyr}'s join functions).}

\item{suffix}{dplyr::join}
}
\value{
data.frame
}
\description{
Join two data.frames using fuzzy string matching.
}
\details{
This function is effectively a customized version of the \code{stringdist_join}
functions in the \code{fuzzyjoin} package.
Additionally, this function is primarily intended to be uzed before \code{tetidy::summarise_join_stats}
with \code{mode = "full"}.
}
\seealso{
\code{\link[fuzzyjoin:stringdist_join]{fuzzyjoin::stringdist_join()}}
}
