% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-combos.R
\name{rank_unique}
\alias{rank_unique}
\title{\code{dplyr::row_number(dplyr::desc(.))}}
\usage{
rank_unique(x)
}
\arguments{
\item{x}{symbol. Name of column on which to perform operations. Should be numeric.}
}
\description{
Shorthand for \code{dplyr} functions called consecutively.
}
\details{
None.
}
